Parachute is being stupid
Here's the problem - I updated the parachute with double sided normals
and a better appearance overall, but in replacing the obj file, the j3o
file is not changing too. So when I ran the game, the parachute is still
having issues. Going to try to figure this out tomorrow